fixup import script for new field
It looks like the campaign finance report data format got updated and
the old script no-longer works. We're a little tied to the "raw" format
so I found the fields that I could match and nullified fields that don't
exist. There are a lot of blank fields in the new report unfortunately.